Change GetAverage return type to double

+++ SimpleConsoleApp/SimpleConsoleApp/StatisticsHelper.cs
@@ -2,14 +2,14 @@
 
 public class StatisticsHelper
 {
-	public static int GetAverage()
+	public static double GetAverage()
 	{
 		Console.WriteLine("Enter some nums to find the average:");
 
 		string userNumsString = Console.ReadLine();
 		string[] userNums = userNumsString.Split();
-		List<int> nums = new List<int>();
-		foreach (string n in userNums) nums.Add(int.Parse(n));
-		return (int)nums.Average();
+		List<double> nums = new List<double>();
+		foreach (string n in userNums) nums.Add(double.Parse(n));
+		return (double)nums.Average();
 	}
 }
